Fix some posible Maria DB error and style broken due ; added on eol.
Fixing some ; error many error due that styles was broken.

Removed ; are unncecesary in that parts of eol. Within min css is the right way to write rule by line and without ;

This is the right way to do :)

media queries either do not need that because this way all styles are broken inside navigators, and more so on parts of styles must be broken due that and some parts would see bad.

And fixed for posible Maria error.